iox-#1032 Move some code from errors.hpp to types.hpp
